Moved HederaLogo to WalletDriver and removed now obsolete MetaMask.ts
